Create Notes component
