+++ src/pages/Ready.tsx
@@ -1,24 +1,14 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import type { Note } from '../types';
 import Navbar from '../components/Navbar.tsx';
-import { createNote, readNotes, deleteNote } from '../hooks/api.ts';
-import { saveNoteLocal } from '../hooks/localStorage.ts';
 import Footer from '../components/Footer.tsx';
-import dayjs from 'dayjs';
+import Notes from '../components/Notes.tsx';
+import { createNote } from '../hooks/api.ts';
+import { saveNoteLocal } from '../hooks/localStorage.ts';
 
 function Ready() {
   const navigate = useNavigate();
   const [subject, setSubject] = useState<string>('');
-  const [notes, setNotes] = useState<Note[]>([]);
-
-  useEffect(() => {
-    (async () => {
-      const notes = await readNotes();
-      if (!notes) return;
-      setNotes(notes);
-    })();
-  }, []);
 
   const handleSubjectChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setSubject(e.target.value);
@@ -30,28 +20,6 @@ function Ready() {
     if (!note) return;
     saveNoteLocal(note.id, note.subject, note.content);
     navigate('/go', { state: { id: note.id, subject: note.subject } });
-  };
-
-  const handleHistoryClick = async (note: Note) => {
-    navigate('/go', {
-      state: { id: note.id, subject: note.subject, content: note.content },
-    });
-  };
-
-  const handleDeleteClick = async (id: string) => {
-    const selectedNote = notes.find((note) => note.id === id);
-    if (!selectedNote) return;
-    const selectedNoteUpdatedAt = dayjs(selectedNote.updatedAt).format(
-      'YYYY/MM/DD HH:mm:ss',
-    );
-    const isConfirmed = confirm(
-      `このメモを削除しますか？\n\n${selectedNote.subject}\n\n${selectedNoteUpdatedAt}`,
-    );
-    if (!isConfirmed) return;
-    const deletedNote = await deleteNote(id);
-    if (!deletedNote) return;
-    const newNotes = notes.filter((note) => note.id !== id);
-    setNotes(newNotes);
   };
 
   return (
@@ -92,33 +60,7 @@ function Ready() {
           <span className="i-ph-bookmarks-simple-thin text-4xl" />
           保存したメモ
         </h2>
-        <div className="mb-6 mx-auto max-w-screen-md">
-          {notes.map((note) => (
-            <div key={note.id} className="flex mb-3 w-full">
-              <button
-                type={'button'}
-                className="py-3 px-4 w-full flex items-center justify-between gap-2 truncate
-                rounded-s-lg border border-gray-200
-                hover:bg-stone-200 focus:bg-stone-200"
-                onClick={() => handleHistoryClick(note)}
-              >
-                <span className="text-base text-gray-800 truncate">
-                  {note.subject}
-                </span>
-                <span className="text-sm text-gray-400">{note.updatedAt}</span>
-              </button>
-              <button
-                type={'button'}
-                onClick={() => handleDeleteClick(note.id)}
-                className="py-3 px-4 -ms-px rounded-e-lg
-                border border-gray-200
-                hover:bg-stone-200 focus:bg-stone-200"
-              >
-                <span className="i-ph-trash-light" />
-              </button>
-            </div>
-          ))}
-        </div>
+        <Notes />
       </div>
       <Footer />
     </>
